+++ src/main/java/com/projects/voting_system/repos/VoteRepos.java
@@ -3,5 +3,5 @@ package com.projects.voting_system.repos;
 import com.projects.voting_system.entities.Vote;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VoteRepos extends JpaRepository<Vote, Integer> {
+public interface VoteRepos extends JpaRepository<Vote, Long> {
 }
